requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.7
django-admin-sortable2==0.7.2
django-appconf==1.0.3
django-easy-select2==1.5.6
django-richtextfield==1.4.0
django-select2==7.0.5
django-social-auth==0.7.28
django-social-login==0.2.0
httplib2==0.13.0
idna==2.8
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
oauth2==1.9.0.post1
oauthlib==3.0.1
Pillow==6.0.0
protobuf==3.8.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
socialoauth==0.3.3
urllib3==1.25.3
